+++ src/main/webapp/css/paginas/conta.css
@@ -1,4 +1,3 @@
-@import url(/css/style.css);
 @import url(/css/crud_cliente.css);
 
 #container-conta {
