+++ src/main/webapp/paginas/conta.html
@@ -1,0 +1,73 @@
+<title>Conta</title>
+
+<nav id="main-nav"></nav>
+
+<h3 class="text-center">Minha Conta</h3>
+
+<div id="container-conta" class="py-5">
+    <div id="nav-lateral" class="d-flex flex-column align-items-start " style="height: 100vh; width: 250px;">
+        <button id="trocar-para-dados-pessoais" type="button" class="btn btn-primary mb-2" onclick="trocarSecao('secao-dados-pessoais')">
+            <img src="/img/imagem-do-usuario-com-fundo-preto.png">
+            Dados Pessoais
+        </button>
+        <button id="trocar-para-endereco" type="button" class="btn btn-success mb-2" onclick="trocarSecao('secao-endereco')">
+            <img src="/img/house-fill.svg">
+            Endereços
+        </button>
+        <button id="trocar-para-cartao" type="button" class="btn btn-secondary mb-2" onclick="trocarSecao('secao-cartao-credito')">
+            <img src="/img/credit-card.svg">
+            Cartões de Crédito
+        </button>
+
+        <button id="trocar-para-cupom" type="button" class="btn btn-warning mb-2" onclick="trocarSecao('secao-cupons')">
+            <img src="/img/ticket.svg" alt="Cupons">
+            Cupons
+        </button>
+    </div>
+
+    <div id="secoes-conta">
+
+        <section id="secao-dados-pessoais" style="display: flex;">
+            <div class="container"></div>
+
+            <div>
+                <button id="btn-editar-cadastro" type="button" class="btn btn-primary btn-sm btn-primary btn-lg" data-bs-toggle="modal"
+                    data-bs-target="#modal-alterar-dados-pessoais">
+                    Editar Cadastro
+                </button>
+                <button id="btn-alterar-senha" type="button" class="btn btn-primary btn-sm btn-primary btn-lg" data-bs-toggle="modal"
+                    data-bs-target="#modal-alterar-senha">
+                    Alterar Senha
+                </button>
+            </div>
+        </section>
+
+        <section id="secao-endereco" style="display: none;">
+            <div class="container-endereco"></div>
+            <div>
+                <button id="btn-adicionar-endereco" type="button" class="btn btn-primary btn-sm btn-primary btn-lg" data-bs-toggle="modal"
+                data-bs-target="#modal-adicionar-endereco">
+                    Adicionar Endereço
+                </button>
+            </div>
+        </section>
+
+        <section id="secao-cartao-credito" style="display: none;">
+            <div class="container-cartoes"></div>
+
+            <div>
+                <button id="btn-adicionar-cartao" type="button" class="btn btn-primary btn-sm btn-primary btn-lg" data-bs-toggle="modal"
+                data-bs-target="#modal-adicionar-cartao-credito">
+                    Adicionar Cartão de crédito
+                </button>
+            </div>       
+        </section>
+
+        <section id="secao-cupons" style="display: none;">
+            <div class="container-cupom"></div>
+        </section>
+
+    </div>
+</div>
+
+<footer id="main-footer"></footer>
